fix: Suspense and fallback, enhance Layout with loading spinner on Layout
